feat: retain newlines in string jsons
This feature will automatically detect whether or not the
existing json string ends in a newline and if it does - then
when stringifying the json, it will append a newline character.

This is useful for occasions where, for example, a developer
desires to have trailing newlines; we want to maintain that
preference.