feat: Keep the "pre" when sorting certain scripts (#50)
* feat: Keep the "pre" when sorting certain scripts

Before, "prepare" was being sorted as if it was named "pare" and
"prettier" was being sorted as if it was named "ttier". These scripts
are now sorted as is. This is done via a blacklist, so other similar
scripts could be added later.

* feat: Only omit "pre" or "post" based on a whitelist

The whitelist contains the default scripts defined in
https://docs.npmjs.com/misc/scripts plus any scripts in the package.json
that appear as a suffix of another script in the same package.json.